requirements.txt
fastapi==0.115.0
pydantic==2.9.2
google-generativeai==0.8.3
uvicorn==0.31.0
jinja2==3.1.4
